backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
deprecation==2.1.0
docopt==0.6.2
fonttools==4.60.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
multidict==6.7.0
nba_api==1.10.2
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pipreq==0.4
pipreqs==0.4.13
postgrest==2.23.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.23.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
storage3==2.23.0
StrEnum==0.4.15
supabase==2.23.0
supabase-auth==2.23.0
supabase-functions==2.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarg==0.1.10
yarl==1.22.0
